requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
apispec==6.4.0
arrow==1.3.0
arxiv==2.1.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.7.0
blis==0.7.11
boto3==1.34.34
botocore==1.34.34
build==1.0.3
cachelib==0.9.0
cachetools==5.3.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
confection==0.1.4
contourpy==1.2.0
cookiecutter==2.5.0
cryptography==42.0.2
curl_cffi==0.6.0b9
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
duckduckgo_search==4.4.3
effdet==0.4.1
emoji==2.10.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.109.2
feedparser==6.0.10
filelock==3.13.1
filetype==1.2.0
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Migrate==4.0.5
flask-smorest==0.42.3
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.27.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
iniconfig==2.0.0
iopath==0.1.10
ipython==8.21.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.5
langchain-community==0.0.19
langchain-core==0.1.22
langchain-experimental==0.0.50
langchain-openai==0.0.5
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.87
layoutparser==0.3.4
limits==3.7.0
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
olefile==0.47
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.11.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
ordered-set==4.1.0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.9.0
pikepdf==8.12.0
pillow==10.2.0
pillow_heif==0.15.0
pip-tools==7.3.0
pluggy==1.4.0
portalocker==2.8.2
posthog==3.4.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
PyJWT==2.8.0
pypandoc==1.12
pyparsing==3.1.1
pypdf==4.0.1
pypdfium2==4.26.0
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.0.0
pytesseract==0.3.10
pytest==8.0.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.7
python-pptx==0.6.23
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
rapidfuzz==3.6.1
redis==5.0.1
regex==2023.12.25
repath==0.9.0
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.7.0
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
srsly==2.4.8
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
text-unidecode==1.3
textblob==0.17.1
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tokenizers==0.15.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
unstructured==0.12.4
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==2.0.7
uvicorn==0.27.0.post1
uvloop==0.19.0
validate-email-address==1
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
webargs==8.4.0
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.16.0
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.4
zipp==3.17.0
